feat: planning optimization (#2)
* feat: bench

* feat: two slices for open and closed nodes

* fix: correct fetchNode on closedNodes

* feat: reduce cost of allowedRepetition calls

If the action is repeatable, we can immediately return true

* chore: clean code

* feat: small improvement on data hash

* feat(astar): use pools to store nodes in memory

* feat(astar): store nodes pointers

* feat: reduce memory usage using float32 for costs

+++ planer_test.go
@@ -6,7 +6,7 @@ func TestPlan_GetTotalCost(t *testing.T) {
 	tests := []struct {
 		name string
 		plan Plan
-		want float64
+		want float32
 	}{
 		{"plan 1", Plan{}, 0.0},
 		{"plan 2", Plan{{
